requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
azure-ai-documentintelligence
azure-storage-blob
azure-search-documents
chonkie
pypdf
openai
Werkzeug